Classifying the crypto namespace.
