+++ src/Hyjinx.LibHac/Crypto/Aes.TPM.cs
@@ -1,4 +1,4 @@
-﻿#if IS_LEGACY_ENABLED
+﻿#if IS_TPM_BYPASS_ENABLED
 
 using LibHac.Common;
 using LibHac.Crypto.Impl;
@@ -10,6 +10,40 @@ namespace LibHac.Crypto;
 
 static partial class Aes
 {
+    public const int KeySize128 = 0x10;
+    public const int BlockSize = 0x10;
+
+    public static bool IsAesNiSupported()
+    {
+        return AesNi.IsSupported;
+    }
+    
+    public static ICipher CreateCbcDecryptor(ReadOnlySpan<byte> key, ReadOnlySpan<byte> iv, bool preferDotNetCrypto = false)
+    {
+        if (IsAesNiSupported() && !preferDotNetCrypto)
+        {
+            return new AesCbcDecryptorNi(key, iv);
+        }
+
+        return new AesCbcDecryptor(key, iv);
+    }
+
+    public static int DecryptCbc128(ReadOnlySpan<byte> input, Span<byte> output, ReadOnlySpan<byte> key,
+        ReadOnlySpan<byte> iv, bool preferDotNetCrypto = false)
+    {
+        if (IsAesNiSupported() && !preferDotNetCrypto)
+        {
+            Unsafe.SkipInit(out AesCbcModeNi cipherNi);
+
+            cipherNi.Initialize(key, iv, true);
+            return cipherNi.Decrypt(input, output);
+        }
+
+        ICipher cipher = CreateCbcDecryptor(key, iv, preferDotNetCrypto);
+
+        return cipher.Transform(input, output);
+    }
+
     public static ICipher CreateEcbDecryptor(ReadOnlySpan<byte> key, bool preferDotNetCrypto = false)
     {
         if (IsAesNiSupported() && !preferDotNetCrypto)
