+++ src/Hyjinx.LibHac/Crypto/ICipher.Legacy.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿#if IS_LEGACY_ENABLED
+
+using System;
 
 namespace LibHac.Crypto;
 
@@ -6,3 +8,5 @@ internal interface ICipher
 {
     int Transform(ReadOnlySpan<byte> input, Span<byte> output);
 }
+
+#endif
