+++ src/Hyjinx.LibHac/Tools/Fs/XciHeader.Legacy.cs
@@ -1,0 +1,12 @@
+﻿#if IS_LEGACY_ENABLED
+
+namespace LibHac.Tools.Fs;
+
+public enum CardClockRate
+{
+    ClockRate25 = 0xA10011,
+    ClockRate50 = 0xA10010
+}
+
+
+#endif
