+++ src/Hyjinx.LibHac/Tools/Fs/XciHeader.cs
@@ -201,12 +201,6 @@ public partial class XciHeader
     }
 }
 
-public enum CardClockRate
-{
-    ClockRate25 = 0xA10011,
-    ClockRate50 = 0xA10010
-}
-
 public enum XciPartitionType
 {
     Update,
